refactor setting API
