+++ .env.example
@@ -20,3 +20,4 @@ JWT_SECRET_KEY=xxxx
 # AWS S3
 S3_REGION=ap-northeast-2
 S3_BUCKET_NAME=poapper-xxxxxxxx
+S3_CF_DIST_URL=xxxx.poapper.club
