+++ src/file/file.service.ts
@@ -4,6 +4,7 @@ import {
   S3Client,
 } from '@aws-sdk/client-s3';
 import { Injectable } from '@nestjs/common';
+import { Readable } from 'stream';
 
 @Injectable()
 export class FileService {
@@ -14,6 +15,26 @@ export class FileService {
   private readonly cfDistUrl: string = process.env.S3_CF_DIST_URL;
 
   constructor() {}
+
+  private streamToString(stream: Readable): Promise<string> {
+    return new Promise((resolve, reject) => {
+      const chunks: Uint8Array[] = [];
+      stream.on('data', (chunk) => chunks.push(chunk));
+      stream.on('error', reject);
+      stream.on('end', () => resolve(Buffer.concat(chunks).toString('utf-8')));
+    });
+  }
+
+  async uploadText(key: string, text: string) {
+    await this.s3.send(
+      new PutObjectCommand({
+        Bucket: this.bucket,
+        Key: key,
+        Body: text,
+      }),
+    );
+    return `${this.cfDistUrl}/${key}`;
+  }
 
   async uploadFile(key: string, file) {
     await this.s3.send(
@@ -27,13 +48,12 @@ export class FileService {
     return `${this.cfDistUrl}/${key}`;
   }
 
-  async deleteFile(key) {
-    const res = await this.s3.send(
+  deleteFile(key: string) {
+    return this.s3.send(
       new DeleteObjectCommand({
         Bucket: this.bucket,
         Key: key,
       }),
     );
-    console.log(res);
   }
 }
