+++ src/popo/setting/setting.module.ts
@@ -1,8 +1,9 @@
-import { CacheModule, Module } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { SettingController } from './setting.controller';
+import { FileModule } from '../../file/file.module';
 
 @Module({
-  imports: [CacheModule.register()],
+  imports: [FileModule],
   controllers: [SettingController],
 })
 export class SettingModule {}
